third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_innerHTML_template-expected.txt

This is a testharness.js-based test.
FAIL html5lib_innerHTML_template.html 64d3e0e4395745b6ea928e5f5cf888bf675a598b assert_equals: expected "#document\n| <template>\n|   content\n|     <form>\n|       <input>\n|         name=\"q\"\n|     <div>\n|       \"second\"" but got "#document\n| content\n|   <template>\n|     content\n|       <form>\n|         <input>\n|           name=\"q\"\n|       <div>\n|         \"second\""
Harness: the test ran to completion.
